+++ src/Classifiers/SoftmaxClassifier.php
@@ -36,6 +36,7 @@ use Generator;
 use function is_nan;
 use function count;
 use function get_object_vars;
+use function number_format;
 
 /**
  * Softmax Classifier
@@ -320,6 +321,10 @@ class SoftmaxClassifier implements Estimator, Learner, Online, Probabilistic, Ve
 
         if ($this->logger) {
             $this->logger->info("Training $this");
+
+            $numParams = number_format($this->network->numParams());
+
+            $this->logger->info("{$numParams} trainable parameters");
         }
 
         $prevLoss = $bestLoss = INF;