+++ tests/NeuralNet/FeedForwardTest.php
@@ -116,6 +116,16 @@ class FeedForwardTest extends TestCase
     /**
      * @test
      */
+    public function numParams() : void
+    {
+        $this->network->initialize();
+
+        $this->assertEquals(103, $this->network->numParams());
+    }
+
+    /**
+     * @test
+     */
     public function roundtrip() : void
     {
         $this->network->initialize();