+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+- 2.4.0
+    - Add GELU activation function
+    - Add numParams() method to Network
+    - Neural Network Learners now report number of trainable parameters
+    - Regex Filter added pattern to match unicode emojis
+    - Custom escape character for CSV Extractor
+    
 - 2.3.4
     - Add string literal type-hints